Improve H_FindPath accuracy via moving to the transpiler system.
If you ever see reports of the pathfinder spiking after pause, check where it is being called from! The pawn tweener (pawn drawing) actually calls some of the pathfinding methods while paused which accumlate and result in a spike. I have no good solution for this, but its an interesting (and un-intended) behavior.